+++ projects/js-packages/charts/src/charts/geo-chart/stories/index.api.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 
-<Meta title="JS Packages/Charts/Types/Geo Chart/API Reference" />
+<Meta title="JS Packages/Charts Library/Charts/Geo Chart/API Reference" />
 
 # Geo Chart API Reference
