+++ projects/js-packages/charts/src/charts/geo-chart/stories/index.docs.mdx
@@ -26,7 +26,7 @@ The Geo Chart component provides a clean, accessible solution for displaying cou
 
 ## API Reference
 
-For detailed information about component props and types, see the [Geo Chart API Reference](./?path=/docs/js-packages-charts-types-geo-chart-api-reference--docs).
+For detailed information about component props and types, see the [Geo Chart API Reference](./?path=/docs/js-packages-charts-library-charts-geo-chart-api-reference--docs).
 
 ## Basic Usage
 
@@ -68,7 +68,7 @@ The simplest geo chart requires data, width, and height. The `data` prop accepts
 - **`scale`**: Scale factor for the map projection. Defaults to `width / (2 * Math.PI)` to fit the map exactly within the chart bounds
 - **`center`**: Geographic center point as `[longitude, latitude]` for zooming into specific regions (e.g., `[-95, 40]` for USA, `[10, 50]` for Europe)
 
-For detailed prop information and type definitions, see the [Geo Chart API Reference](./?path=/docs/js-packages-charts-types-geo-chart-api-reference--docs).
+For detailed prop information and type definitions, see the [Geo Chart API Reference](./?path=/docs/js-packages-charts-library-charts-geo-chart-api-reference--docs).
 
 ## Interactive Features
